requirements.txt
google-genai>=0.2.0
python-dotenv>=1.0.0
python-docx>=1.0.0
PyPDF2>=3.0.0
pandas>=2.0.0
python-pptx>=0.6.0
